Simplify auth pages
