+++ frontend/src/app/(auth)/home/pages/Dashboard.tsx
