+++ frontend/src/app/layout.tsx
@@ -1,16 +1,13 @@
-import type { Metadata } from "next";
+"use client";
+
 import { Inter } from "next/font/google";
 import "./globals.css";
 import { GoogleOAuthProvider } from "@react-oauth/google";
+import { AuthProvider } from "@/components/auth/AuthContext";
 
 const inter = Inter({
   subsets: ["latin"],
 });
-
-export const metadata: Metadata = {
-  title: "PeerPrep",
-  description: "A CS3219 Project by Group 15!",
-};
 
 export default function RootLayout({
   children,
@@ -21,7 +18,9 @@ export default function RootLayout({
     <html lang="en" className={`${inter.className} h-full`}>
       <body className="h-full">
         <GoogleOAuthProvider clientId="785838083864-7jmrr23k3homjemh5n2fvk7ouk759eb6.apps.googleusercontent.com">
-          {children}
+          <AuthProvider>
+            {children}
+          </AuthProvider>
         </GoogleOAuthProvider>
       </body>
     </html>
